Refactor: Move the WorldServer to a separate module
